requirements.txt
Django==2.2.7
django-allauth==0.40.0
djangorestframework===3.10.3
